Revert "Main page fixes (#42)"
This reverts commit c7973d2e066687c309a433ee8e69e449d75bb5ab.
