+++ src/pages/index.tsx
@@ -11,8 +11,9 @@ import { trpc } from "../utils/trpc";
 import DropDownMenu from "../components/DropDownMenu";
 import { browserEnv } from "../utils/env/browser";
 import ProtectedPage from "../utils/auth";
+import Sidebar from "../components/Sidebar";
 import Header, { HeaderOptions } from "../components/Header";
-import { PublicUser, User } from "../utils/types";
+import { PublicUser } from "../utils/types";
 import ConnectModal from "../components/ConnectModal";
 import { toast } from "react-toastify";
 import ExploreSidebar from "../components/ExploreSidebar";
@@ -31,13 +32,8 @@ const Home: NextPage<any> = () => {
     isLoading: isLoadingUser,
     refetch,
   } = trpc.useQuery(["user.me"]);
-
   const { data: recommendations } = trpc.useQuery(["user.recommendations.me"]);
   const { data: favorites } = trpc.useQuery(["user.favorites.me"]);
-  const { data: requests } = trpc.useQuery(["user.requests.me"]);
-  const fromRequestIds =
-    requests?.from.map((request) => request.fromUserId) ?? [];
-  const toRequestIds = requests?.to.map((request) => request.toUserId) ?? [];
   const { mutate: mutateFavorites } = trpc.useMutation("user.favorites.edit", {
     onError: (error) => {
       toast.error(`Something went wrong: ${error.message}`);
@@ -50,59 +46,18 @@ const Home: NextPage<any> = () => {
   const [mapState, setMapState] = useState<mapboxgl.Map>();
 
   const [modalUser, setModalUser] = useState<PublicUser | null>(null);
-  const [modalType, setModalType] = useState<string>("connect");
   const [sidebarState, setSidebarState] = useState<HeaderOptions>("explore");
 
   const handleConnect = (userToConnectTo: PublicUser) => {
     setModalUser(userToConnectTo);
-    setModalType("connect");
-  };
-
-  const { mutate: mutateRequests } = trpc.useMutation("user.requests.create", {
-    onError: (error) => {
-      toast.error(`Something went wrong: ${error.message}`);
-    },
-    onSuccess() {
-      utils.invalidateQueries(["user.requests.me"]);
-    },
-  });
-
-  const handleEmailConnect = (curUser: User, toUser: PublicUser) => {
-    connectEmail(toUser.email);
-    mutateRequests({
-      fromId: curUser.id,
-      toId: toUser.id,
-      message: "Not sure if we need this",
-    });
-  };
-  const connectEmail = async (email: string | null) => {
-    const msg = {
-      to: email, // Replace with your recipient
-      from: "devashishsood9@gmail.com", // Replace with your verified sender
-      subject: "New Contact Message",
-      text: `Dior Dior`,
-    };
-
-    const result = await fetch(`/api/sendEmail`, {
-      method: "POST",
-      body: JSON.stringify(msg),
-    });
-    console.log(result);
   };
 
   const handleSentRequests = (userToConnectTo: PublicUser) => {
     setModalUser(userToConnectTo);
-    setModalType("sent");
   };
 
   const handleReceivedRequests = (userToConnectTo: PublicUser) => {
     setModalUser(userToConnectTo);
-    setModalType("received");
-  };
-
-  const convertToUser = (id: string) => {
-    const { data: curUser } = trpc.useQuery(["user.else", { id: id }]);
-    return curUser;
   };
 
   useEffect(() => {
@@ -153,12 +108,8 @@ const Home: NextPage<any> = () => {
         return (
           <RequestSidebar
             currentUser={user}
-            sent={toRequestIds
-              .map(convertToUser)
-              .filter((user): user is PublicUser => !!user)}
-            received={fromRequestIds
-              .map(convertToUser)
-              .filter((user): user is PublicUser => !!user)}
+            sent={getPublicUserArray("Sent Users")}
+            received={getPublicUserArray("Received Users")}
             favs={favorites ?? []}
             map={mapState}
             handleSent={handleSentRequests}
@@ -195,17 +146,16 @@ const Home: NextPage<any> = () => {
           {/* map wrapper */}
           <div className="relative flex-auto">
             <div id="map" className={"flex-auto w-full h-full"}></div>
-            {user && modalUser && modalType === "connect" && (
+            {user && modalUser && (
               <ConnectModal
                 currentUser={user}
                 userToConnectTo={modalUser}
-                handleEmailConect={() => handleEmailConnect(user, modalUser)}
                 closeModal={() => {
                   setModalUser(null);
                 }}
               />
             )}
-            {user && modalUser && modalType === "sent" && (
+            {user && modalUser && (
               <SentRequestModal
                 currentUser={user}
                 userToConnectTo={modalUser}
@@ -214,7 +164,7 @@ const Home: NextPage<any> = () => {
                 }}
               />
             )}
-            {user && modalUser && modalType === "received" && (
+            {user && modalUser && (
               <ReceivedRequestModal
                 currentUser={user}
                 userToConnectTo={modalUser}
